move lookups to separate file
