+++ controllers/project_mesh_controller.go
@@ -22,8 +22,6 @@ const (
 	AnnotationGateway     = "opendatahub.io/service-mesh-gw"
 	LabelMaistraGw        = "maistra.io/gateway-name"
 	LabelMaistraGwNs      = "maistra.io/gateway-namespace"
-	MeshNamespaceEnv      = "MESH_NAMESPACE"
-	ControlPlaneEnv       = "CONTROL_PLANE_NAME"
 )
 
 // OpenshiftServiceMeshReconciler holds the controller configuration.
